Update patch set 1
Patch Set 1:

To which check are you referring? FileName != NULL or *FileName != 0? The former, I would agree, since that is causing a compile error with Clang. The latter *appears* to be some kind of check to see if FileName was ever modified  (see Line 19). A followup is not out of the question (although I probably won't be pursuing such a thing), but that's decidedly out of the scope of fixing the compile errors with Clang.

Patch-set: 1